changed 'ffmpeg -i' use to 'ffprobe'
'ffmpeg -i' always exits with an error (because there is no output file, ffmpeg has no option to only get file info). So there is no way to handle inputfiles that are no valid videofiles and therefore no easy way to prevent the code from breaking on invalid files.

'ffprobe' only exits with an error if it fails to get the video info and we can handle errors here correctly.